refactor: modularise deep-search financial agent package
