+++ python/agents/deep-search/app/config.py
@@ -35,7 +35,8 @@ else:
     import google.auth
 
     _, project_id = google.auth.default()
-    os.environ.setdefault("GOOGLE_CLOUD_PROJECT", project_id)
+    if project_id:
+        os.environ.setdefault("GOOGLE_CLOUD_PROJECT", project_id)
     os.environ["GOOGLE_CLOUD_LOCATION"] = "global"
     os.environ.setdefault("GOOGLE_GENAI_USE_VERTEXAI", "True")
